requirements.txt
absl-py==0.6.1
astor==0.7.1
certifi==2019.3.9
cffi==1.11.5
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
future==0.17.1
gast==0.2.0
grpcio==1.16.0
gym==0.10.9
h5py==2.8.0
idna==2.7
imageio==2.4.1
joblib==0.13.2
julia==0.2.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.1
mkl-fft==1.0.6
mkl-random==1.0.1
networkx==2.3
numpy==1.16.4
olefile==0.46
Pillow==5.4.1
protobuf==3.6.1
pycparser==2.19
-e git+https://github.com/cdrckrgt/PyFEBOL@14858312f7cbd72cf8830f639fcb8cb83c8c5746#egg=PyFEBOL
pyglet==1.3.2
pyparsing==2.3.0
PyQt5==5.11.3
PyQt5-sip==4.19.13
python-dateutil==2.7.5
pytz==2018.7
PyWavelets==1.0.3
PyYAML==5.1
requests==2.22.0
scikit-image==0.15.0
scikit-learn==0.21.1
scipy==1.1.0
six==1.11.0
tb-nightly==1.14.0a20190528
termcolor==1.1.0
torch==1.1.0
torchsummary==1.5.1
torchvision==0.2.1
tornado==5.1.1
urllib3==1.25.2
Werkzeug==0.14.1
